Release version 0.4.1 (backport)
As explained in #1331 we don't currently have a backport of our
docker builds for v0.4.0. However we do have a :latest docker
image which is going to be the default way for people to try out
the docker image for the first time. That is based on the
08408a9 commit.

However, considering the #1503 security patch for our docker image,
I would like to make sure the :latest tag receives this patch as well.
The most straightforward way is to create a v0.4.1 patch release.