Update README.md
The README explains how to set the SOURCECRED_GITHUB_TOKEN later in the README, but later in the README in the Docker section. People who aren't using Docker will follow the initial installation instructions. For this reason it's important that the instructions to set the environment variable are part of the initial setup process.